refactor: move automatic check options, tests to their own file
